curagenie-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
boto3==1.34.0
biopython==1.83
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
alembic==1.13.1
openai==1.3.7
anthropic==0.8.1
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.2
tensorflow==2.15.0
opencv-python==4.8.1.78
imutils==0.5.4
Pillow==10.1.0
